backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
rq==1.16.1
paddleocr==2.7.3
paddlepaddle==3.2.2
opencv-python==4.6.0.66
pillow==10.2.0
numpy==1.26.3
pandas==2.2.0
openai==1.10.0
pydantic==2.5.3
aiofiles==23.2.1
